(refactor) rename exercise 1 and 2 vue files
